spiced up readme with new display abstraction
