+++ dotmatrix/displays/__init__.py
@@ -1,4 +1,5 @@
 from .block import Block
 from .braille import Braille
+from .unit import Unit
 
-__all__ = ("Braille", "Block")
+__all__ = ("Braille", "Block", "Unit")
